requirements.txt
asgiref==3.8.1
astroid==2.5.2
atomicwrites==1.4.0
attrs==20.1.0
beautifulsoup4>=4.10.0
click==8.1.3
click-completion==0.5.2
colorama==0.4.3
crayons==0.4.0
dj-static==0.0.6
Django==4.2
django-bootstrap4==24.4
django-filter==24.3
django-tables2==2.7.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
docopt==0.6.2
et-xmlfile==1.1.0
factory-boy==3.2.0
Faker==17.6.0
gunicorn==20.0.4
iniconfig==1.0.1
isort==5.4.2
jdcal==1.4.1
Jinja2==3.1.2
lazy-object-proxy==1.6.0
MarkupSafe==2.1.2
mccabe==0.6.1
mock==4.0.2
more-itertools==8.4.0
openpyxl==2.6.0
packaging==20.4
Pillow==9.3.0
pluggy==0.13.1
psycopg2-binary==2.9.5
py==1.9.0
pycodestyle==2.7.0
pycycle==0.0.8
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.0.1
pytest-django==3.8.0
python-dateutil==2.8.2
pytz==2020.1
shellingham==1.5.0.post1
six==1.15.0
sorl-thumbnail==12.7.0
soupsieve==2.0.1
sqlparse==0.3.1
static3==0.7.0
template-remover==0.1.9
toml==0.10.1
typing-extensions==4.12.2
tzdata==2025.1
untokenize==0.1.1
wrapt==1.12.1